Adaptando EP1 para funcionar com o autograder
O coursera deixa submetermos custom autograders como docker containers.
Pra facilitar a submissao e correcao desse trabalho eu agreguei varios
dos utils pra dentro do word2vec.py para poder submeter um arquivo só.

Alem disso apaguei do enunciado a parte de derivadad já que não são tão
interessantes pra nós no contexto do coursera e são  mais dificeis de
validar.

Tb consertei um typo no run.py
